+++ Modules/Intent.Modules.AspNetCore.OData.EntityFramework/release-notes.md
@@ -1,5 +1,7 @@
 ### Version 1.0.0
 
+- Improvement: Added context menu item on aggregates `Expose with OData` as well as limited stereotype usage to only aggregates.
+
 - Fixed: Added `[ODataRouteComponent("odata")]` to all ODataControllers.
 
 - Improvement: Now only checks for Entities that have the IHasDomainEvent implemented.